+++ src/main/java/step2/domain/Lotto.java
@@ -1,4 +1,4 @@
-package step2;
+package step2.domain;
 
 public class Lotto {
