+++ src/main/java/step2/domain/LottoMachine.java
@@ -1,5 +1,6 @@
-package step2;
+package step2.domain;
 
+import step2.util.LottoNumberGenerator;
 import java.util.ArrayList;
 import java.util.List;
