+++ src/test/java/step2/domain/UserTest.java
@@ -1,4 +1,4 @@
-package step2;
+package step2.domain;
 
 import org.junit.jupiter.api.Test;
 import java.util.List;
